requirements.txt
google-genai
mistralai
openpyxl
pandas
fastapi[standard]
uvicorn[standard]
pydantic
httpx
python-multipart
google-auth
